New demo types in new files
