+++ Demo/Demo/Gravatar-Demo/SwiftUI/Controls/AvatarPickerLayoutOptions.swift
