+++ Demo/Demo/Gravatar-Demo/SwiftUI/Controls/QEColorSchemePickerRow.swift
